Version 0.2
- Added support for images over keys
- Added key specific styles

Note from David Wilhelm:
This is a rewritten version of
986600003215729b7e3aad50596e642612dd80a3 from the original repo.

This version of the commit is identical to the original except
that it excludes gotri.exe and the build events that run it.

This is done for security, so NohBoard compiles out of the box
using only source code, without running any black box binaries.

Version info will instead be maintained manually.