Fix bugs in Issue #34
Buttons seemed to randomly disappear after doing things when a new
keyboard definition was just loaded. Then there was a crash after the
program tried to render the selected element but no manipulation was
set. This happens when the properties form is opened, this selects the
element but does not set a manipulation.